Go to the Postgresql

+++ app/goods/models.py
@@ -20,7 +20,7 @@ class Products(models.Model):
     slug = models.SlugField(max_length=200, unique=True, blank=True, null=True, verbose_name="URL")
     description = models.TextField(blank=True, null=True, verbose_name="Описание")
     image = models.ImageField(upload_to="goods_images", blank=True, null=True, verbose_name="Изображение")
-    price = models.PositiveSmallIntegerField(default=0, blank=True, null=True, verbose_name="Цена")
+    price = models.PositiveIntegerField(default=0, blank=True, null=True, verbose_name="Цена")
     discount = models.SmallIntegerField(default=0, blank=True, null=True, verbose_name="Скидка в %")
     quantity = models.PositiveSmallIntegerField(default=0, blank=True, null=True, verbose_name="Количество")
     category = models.ForeignKey(to=Categories, on_delete=models.CASCADE, verbose_name="Категория")
